[feat] 툴팁 수정
